+++ packages/esp_capture/test_apps/main/test_cxx_build.cpp
@@ -31,6 +31,7 @@ extern "C" void test_cxx_build(void)
             .sync_mode = ESP_CAPTURE_SYNC_MODE_AUDIO,
             .audio_src = audio_src,
             .video_src = video_src,
+            .share_overlay = false,
         };
         esp_capture_open(&capture_cfg, &capture);
         esp_capture_sink_cfg_t sink_cfg = {